app/requirements.txt
fastapi==0.115.7
uvicorn==0.34.0
sqlalchemy==2.0.37
pydantic==2.10.6
python-multipart==0.0.20
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
joblib==1.3.2
tiktoken==0.6.0
openai
chromadb==0.4.22
tqdm==4.66.2
python-dotenv==1.0.1
aiofiles==23.2.1
